revert mgmt command names so changes have a reviewable diff
